refactor: - app router next-auth 설정 삭제 - page router next-auth 설정으로 변경 - build 오류 수정
